Add Breadcrumbs and Details page edit
Created breadcrumbs for page changes. Due to one database and page limit
the product type in breadcrumb directs the customer to the
index.php(product list page). Added Functions to only display the sale
percentage on the item which has a sale or which sale price is not
higher than the actual price. Added php statement in details page to
display only the products price if the sale is equal to zero and put a
line through the maine price if there is a sale.
removed header and footer additional pages.